Add relative strength to graphs
If a user inputs new data and their body weight is
already recorded, the weight is linked to the data.
This allows us to plot the user’s strength against
their weight at that time. For instance, you might
see an increase in a specific lift, but when adjusted
for body weight, it could actually be a total mass
increase.